+++ feature/challenges/src/main/kotlin/uk/co/weightwars/challenges/SharedChallengeViewModel.kt
@@ -6,15 +6,15 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SharedChallengeViewModel @Inject constructor() : ViewModel() {
-    private var challengeId: Long? = null
+    private var challengeId: Int? = null
 
-    fun getChallenge(): Long? {
+    fun getChallenge(): Int? {
         val challenge = this.challengeId
         this.challengeId = null
         return challenge
     }
 
-    fun setChallenge(challengeId: Long) {
+    fun setChallenge(challengeId: Int) {
         this.challengeId = challengeId
     }
 }
