+++ feature/challenges/src/main/kotlin/uk/co/weightwars/challenges/challenge/ChallengeViewModel.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
+import uk.co.weightwars.data.models.Challenge
 import uk.co.weightwars.data.repository.ChallengeRepo
-import uk.co.weightwars.database.entities.Challenge
 import javax.inject.Inject
 
 data class ChallengeUiState(val challenges: List<Challenge> = listOf())
@@ -19,7 +19,7 @@ class ChallengeViewModel @Inject constructor(
 ) : ViewModel() {
     var uiState by mutableStateOf(ChallengeUiState())
 
-    fun getChallenges(categoryId: Long) = viewModelScope.launch {
+    fun getChallenges(categoryId: Int) = viewModelScope.launch {
         val challenges = challengeRepo.getChallenges(categoryId)
         uiState = uiState.copy(challenges = challenges)
     }
